+++ packages/patternfly-4/react-core/src/experimental/components/DataToolbar/DataToolbarUtils.tsx
@@ -6,27 +6,15 @@ interface DataToolbarContextProps {
   toggleIsExpanded: () => void;
   expandableContentRef: RefObject<HTMLDivElement>;
   expandableContentId: string;
+  chipGroupContentRef: RefObject<HTMLDivElement>;
 }
 
 export const DataToolbarContext = React.createContext<Partial<DataToolbarContextProps>>({});
 
 export type DataToolbarBreakpointMod = {
   /** The attribute to modify  */
-  modifier: 'hidden' | 'visible' | 'align-right' | 'align-left';
+  modifier: 'hidden' | 'visible' | 'align-right' | 'align-left' | 'spacer-none' | 'spacer-sm' | 'spacer-md' |
+    'spacer-lg' | 'space-items-none' | 'space-items-sm' | 'space-items-md' | 'space-items-lg';
   /** The breakpoint at which to apply the modifier */
   breakpoint: 'md' | 'lg' | 'xl' | '2xl';
-};
-
-export type DataToolbarSpacer = {
-  /** The size of the spacer */
-  spacerSize: 'none' | 'sm' | 'md' | 'lg';
-  /** The breakpoint at which to apply the spacer */
-  breakpoint?: 'md' | 'lg' | 'xl';
-};
-
-export const formatSpacers = (spacers: DataToolbarSpacer[], type = 'pf-m-spacer') => {
-
-  return spacers.reduce((acc, curr) => (
-    `${acc} ${type}-${curr.spacerSize}${curr.breakpoint ? `-on-${curr.breakpoint}` : ''}`
-  ), '');
 };