+++ packages/patternfly-4/react-topology/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.8.62](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@2.8.61...@patternfly/react-topology@2.8.62) (2019-09-27)
+
+**Note:** Version bump only for package @patternfly/react-topology
+
+
+
+
+
 ## [2.8.61](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@2.8.60...@patternfly/react-topology@2.8.61) (2019-09-26)
 
 **Note:** Version bump only for package @patternfly/react-topology